chore: add useCourses hook
